Properly show date last updated
This is now explicitly defined, and is listed in the footer. It's also
shoehorned into the header, as the acl2015 class does not directly
support the \date directive, ignoring it in its \maketitle title
instructions. It can't be included in the author area as the author area
is in a tabular environment, and it can't be included after the title as
the document enters two-column mode at that point.

Instead, it sits immediately below the title in normal font, unbolded
text.